Quality control and top peak selection
Code for additional quality control plots and linear regression fit to gene TPM vs # reads in peak after removal of peaks with <20 reads. Peaks are sorted according to their residual and the top 1000 peaks are then selected. They are very similar to the top 1000 peaks based on score, so the fit might not be necessary.